refactor convert function implement :  * tab  * long press  * moveCamera  * getVisibleRegion  * getScreenCoordinate  * getLatLng  * getZoomLevel
