+++ lib/google_maps_flutter_web.dart
@@ -6,6 +6,7 @@ import 'dart:ui' as ui;
 import 'dart:math' show sqrt;
 import 'dart:collection';
 import 'dart:convert';
+import 'dart:typed_data';
 
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
@@ -22,7 +23,7 @@ import 'package:google_maps/google_maps.dart' as GoogleMap;
 //import 'package:js/js.dart';
 
 part 'src/google_maps_flutter_web.dart';
-part 'src/google_map_web.dart';
+part 'src/google_maps_controller.dart';
 part 'src/circle.dart';
 part 'src/circles.dart';
 part 'src/polygon.dart';
@@ -31,4 +32,4 @@ part 'src/polyline.dart';
 part 'src/polylines.dart';
 part 'src/marker.dart';
 part 'src/markers.dart';
-part 'src/options.dart';
+part 'src/convert.dart';
