--- lib/src/options.dart
@@ -1,28 +1,0 @@
-part of google_maps_flutter_web;
-
-void _optionsFromParams(GoogleMap.MapOptions options,
-    Map<String, dynamic> optionsUpdate) {
-      print('>'+ optionsUpdate.toString());
-//      compassEnabled: true
-//      mapToolbarEnabled: true
-//      cameraTargetBounds: [null]
-//      mapType: 1
-//  ..mapTypeId(optionsUpdate['mapType'])
-//      minMaxZoomPreference: [null
-//        null]
-//      rotateGesturesEnabled: true
-//      scrollGesturesEnabled: true
-//      tiltGesturesEnabled: true
-//      zoomGesturesEnabled: true
-//      trackCameraPosition: true
-//      myLocationEnabled: true
-//      myLocationButtonEnabled: true
-//      padding: [0
-//        0
-//        0
-//        0]
-//      indoorEnabled: true
-//      trafficEnabled: false
-//      buildingsEnabled: true
-      ;
-}
