+++ lib/src/polyline.dart
@@ -1,60 +1,5 @@
 part of google_maps_flutter_web;
 
-PolylineUpdates _polylineFromParams(value) {
-  if (value != null) {
-    List<Map<String, dynamic>> list = value;
-    Set<Polyline> current = Set<Polyline>();
-    list.forEach((polyline) {
-      PolylineId polylineId = PolylineId(polyline['polylineId']);
-      List<LatLng> points = [];
-      List<dynamic> jsonPoints = polyline['points'];
-      jsonPoints.forEach((p) {
-        points.add(LatLng.fromJson(p));
-      });
-      current.add(
-          Polyline(
-            polylineId: polylineId,
-            consumeTapEvents: polyline['consumeTapEvents'],
-            color: Color(polyline['color']),
-            geodesic: polyline['geodesic'],
-            visible: polyline['visible'],
-            zIndex: polyline['zIndex'],
-            width: polyline['width'],
-            points: points,
-//            endCap = Cap.buttCap,
-//            jointType = JointType.mitered,
-//            patterns = const <PatternItem>[],
-//            startCap = Cap.buttCap,
-          )
-      );
-    });
-    return PolylineUpdates.from(null, current);
-  }
-  return null;
-}
-
-GoogleMap.PolylineOptions _polylineOptionsFromPolyline(GoogleMap.GMap googleMap,
-    Polyline polyline) {
-  List<GoogleMap.LatLng> paths = [];
-  polyline.points.forEach((point) {
-    paths.add(GoogleMap.LatLng(point.latitude, point.longitude));
-  });
-
-  return GoogleMap.PolylineOptions()
-    ..path = paths
-    ..strokeOpacity = 1.0
-    ..strokeWeight = polyline.width
-    ..strokeColor =  '#'+polyline.color.value.toRadixString(16).substring(0,6)
-    ..visible = polyline.visible
-    ..zIndex = polyline.zIndex
-    ..geodesic = polyline.geodesic
-  ;
-//  this.endCap = Cap.buttCap,
-//  this.jointType = JointType.mitered,
-//  this.patterns = const <PatternItem>[],
-//  this.startCap = Cap.buttCap,
-//  this.width = 10,
-}
 
 class PolylineController {
